refactor: rename get-cache-filename to get-project-data-key and implement to project data service. implement new ProjectFilter type and ProjectFilterUtil class
